Split Alacritty binary and library
This is an alternative proposal for
https://github.com/jwilm/alacritty/pull/2312 which splits the main
Alacritty entry point and the Alacritty terminal code into two separate
modules.

Currently this just adds a separate `alacritty_terminal` crate inside
the root project, however further separation should be possible.
