+++ Cargo.toml
@@ -1,92 +1,41 @@
 [package]
 name = "alacritty"
 version = "0.3.0"
-authors = ["Joe Wilm <joe@jwilm.com>"]
 license = "Apache-2.0"
-build = "build.rs"
-description = "GPU-accelerated terminal emulator"
-readme = "README.md"
-homepage = "https://github.com/jwilm/alacritty"
 edition = "2018"
+description = "GPU-accelerated terminal emulator"
+
+readme = "README.md"
+authors = ["Joe Wilm <joe@jwilm.com>"]
+repository = "https://github.com/jwilm/alacritty"
+categories = ["emulators", "text-processing"]
+keywords = ["alacritty", "terminal", "term", "emulator", "opengl"]
 
 [workspace]
 members = [
-    "font",
+    "alacritty_terminal",
     "copypasta",
-    "winpty"
+    "winpty",
+    "font",
 ]
 
-[[bin]]
-doc = false
-path = "src/main.rs"
-name = "alacritty"
-
 [dependencies]
-libc = "0.2"
-notify = "4"
-bitflags = "1"
-font = { path = "./font" }
-errno = "0.2"
-parking_lot = "0.7"
-serde = "1"
-serde_derive = "1"
-serde_json = "1"
-serde_yaml = "0.8"
-vte = "0.3"
-mio = "0.6"
-mio-extras = "2"
-copypasta = { path = "./copypasta" }
-xdg = "2"
-log = "0.4"
-clap = "2"
-fnv = "1"
-unicode-width = "0.1"
-glutin = { version = "0.21.0-rc2", features = ["icon_loading"] }
-env_logger = "0.6.0"
-base64 = "0.10.0"
-static_assertions = "0.3.0"
-terminfo = "0.6.1"
-url = "1.7.1"
-time = "0.1.40"
+alacritty_terminal = { path = "./alacritty_terminal" }
 crossbeam-channel = "0.3.8"
+env_logger = "0.6.0"
+time = "0.1.42"
+log = "0.4.6"
 
-[target.'cfg(unix)'.dependencies]
-nix = "0.13"
-signal-hook = { version = "0.1", features = ["mio-support"] }
-
-[target.'cfg(any(target_os = "linux", target_os = "freebsd", target_os="dragonfly", target_os="openbsd"))'.dependencies]
-x11-dl = "2"
-
-[target.'cfg(windows)'.dependencies]
-winpty = { path = "./winpty" }
-mio-named-pipes = "0.1"
-miow = "0.3"
-dunce = "1.0"
-winapi = { version = "0.3.7", features = ["impl-default", "winuser", "synchapi", "roerrorapi", "winerror", "wincon", "wincontypes"]}
-dirs = "1.0"
-widestring = "0.4"
-mio-anonymous-pipes = "0.1"
-image = "0.21.0"
-
-[target.'cfg(target_os = "macos")'.dependencies]
-objc = "0.2.2"
-dirs = "1.0.2"
+[dev-dependencies]
+serde = "1"
+serde_json = "1"
+serde_derive = "1"
 
 [features]
 default = []
 # Enabling this feature makes shaders automatically reload when changed
-live-shader-reload = []
-nightly = []
+live-shader-reload = ["alacritty_terminal/live-shader-reload"]
 bench = []
-
-[build-dependencies]
-gl_generator = "0.11.0"
-
-[target.'cfg(windows)'.build-dependencies]
-embed-resource = "1.1.4"
-tempfile = "3.0.4"
-reqwest = "0.9"
-zip = "0.5"
 
 [profile.release]
 lto = true
