+++ alacritty_terminal/src/cli.rs
