+++ alacritty_terminal/src/config/bindings.rs
