+++ alacritty_terminal/src/display.rs
