+++ alacritty_terminal/src/grid/storage.rs
