+++ alacritty_terminal/src/index.rs
