+++ alacritty_terminal/src/lib.rs
@@ -14,7 +14,6 @@
 //
 //! Alacritty - The GPU Enhanced Terminal
 #![deny(clippy::all, clippy::if_not_else, clippy::enum_glob_use, clippy::wrong_pub_self_convention)]
-#![cfg_attr(feature = "nightly", feature(core_intrinsics))]
 #![cfg_attr(all(test, feature = "bench"), feature(test))]
 
 #[macro_use]
@@ -39,7 +38,6 @@ pub mod grid;
 pub mod index;
 pub mod input;
 pub mod locale;
-pub mod logging;
 pub mod message_bar;
 pub mod meter;
 pub mod panic;
