+++ alacritty_terminal/src/locale.rs
