+++ alacritty_terminal/src/macros.rs
