+++ alacritty_terminal/src/message_bar.rs
