+++ alacritty_terminal/src/meter.rs
@@ -20,7 +20,7 @@
 //!
 //! ```rust
 //! // create a meter
-//! let mut meter = alacritty::meter::Meter::new();
+//! let mut meter = alacritty_terminal::meter::Meter::new();
 //!
 //! // Sample something.
 //! {
