+++ alacritty_terminal/src/panic.rs
