+++ alacritty_terminal/src/renderer/rects.rs
