+++ alacritty_terminal/src/selection.rs
