+++ alacritty_terminal/src/sync.rs
