+++ alacritty_terminal/src/term/cell.rs
