+++ alacritty_terminal/src/url.rs
